start removing custom selection
