+++ src/napari/utils/events/__init__.py
@@ -4,15 +4,13 @@ from napari.utils.events.event import (  # isort:skip
     EventEmitter,
     set_event_tracing_enabled,
 )
-from napari.utils.events.containers._evented_dict import EventedDict
-from napari.utils.events.containers._evented_list import EventedList
-from napari.utils.events.containers._nested_list import NestableEventedList
-from napari.utils.events.containers._selectable_list import (
+from napari.utils.events.containers import (
+    EventedDict,
+    EventedList,
+    EventedSet,
     SelectableEventedList,
+    Selection,
 )
-from napari.utils.events.containers._selection import Selection
-from napari.utils.events.containers._set import EventedSet
-from napari.utils.events.containers._typed import TypedMutableSequence
 from napari.utils.events.event_utils import disconnect_events
 from napari.utils.events.evented_model import EventedModel
 from napari.utils.events.types import SupportsEvents
@@ -25,11 +23,9 @@ __all__ = [
     'EventedList',
     'EventedModel',
     'EventedSet',
-    'NestableEventedList',
     'SelectableEventedList',
     'Selection',
     'SupportsEvents',
-    'TypedMutableSequence',
     'disconnect_events',
     'set_event_tracing_enabled',
 ]
